Match CGItemObj onNewFinished

+++ src/itemobj.cpp
@@ -1294,7 +1294,7 @@ void CGItemObj::loadModel()
  */
 void CGItemObj::onNewFinished()
 {
-	*(u32*)((u8*)this + 0x568) = *(u32*)((u8*)this + 0x144);
+	*(float*)((u8*)this + 0x568) = *(float*)((u8*)this + 0x144);
 	*(u16*)((u8*)this + 0x560) = (u16)((gItemObjCreateFlags >> 3) & 1);
 	loadModel();
 }
